added page subfolders, combined small, unreused components into page component
