+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import LandingPage from './components/LandingPage';
-import Dashboard from './pages/Dashboard';
+import LandingPage from './pages/Landing/Landing';
+import Dashboard from './pages/Dashboard/Dashboard';
 import { AuthManager } from './components/AuthManager';
 
 import './App.css';
