+++ frontend/src/components/AuthManager.tsx
@@ -3,7 +3,7 @@ import { GoogleLogin } from 'react-google-login';
 import { useHistory, useLocation } from 'react-router-dom';
 import useClientId from '../hooks/useClientId';
 import AuthContext from '../context/auth';
-import LandingPage from './LandingPage';
+import LandingPage from '../pages/Landing/Landing';
 
 export const AuthManager: FunctionComponent = ({ children }) => {
   const [signedIn, setSignedIn] = useState(false);
