--- frontend/src/components/footer.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import '../styles/footer.css';
-
-const Footer = () => (
-  <footer className="footer">
-    Some information about copyright can go here
-  </footer>
-);
-
-export default Footer;
