--- frontend/src/components/header.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import '../styles/header.css';
-
-const Header = () => (
-  <header className="header">
-    <h1 className="title">Welcome to Carriage</h1>
-    <h2 className="subtitle">Helping People Do Something</h2>
-  </header>
-);
-
-export default Header;
