--- frontend/src/components/readmore.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import '../styles/readmore.css';
-
-const ReadMore = () => (
-  <a href="https://www.cornelldti.org" className="read-btn">
-    Read More
-  </a>
-);
-
-export default ReadMore;
