+++ frontend/src/pages/Landing/landing.module.css
@@ -1,0 +1,38 @@
+.home {
+  background-image: url("../img/showcase.jpg");
+  background-size: cover;
+  background-size: center;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  padding: 0 20px;
+}
+
+.title, .subtitle {
+  margin: 0
+}
+
+.footer {
+  padding: 1rem;
+  background: black;
+  color: white;
+  text-align: center;
+}
+
+.read-btn {
+  font-size: 1.1em;
+  text-decoration: none;
+  color: black;
+  border: black 1px solid;
+  padding: 10px 20px;
+  border-radius: 10px;
+  margin-top: 10px;
+}
+
+.read-btn:hover {
+  background: black;
+  color: white;
+}
