--- frontend/src/styles/footer.css
@@ -1,6 +1,0 @@
-.footer {
-  padding: 1rem;
-  background: black;
-  color: white;
-  text-align: center;
-}
