--- frontend/src/styles/header.css
@@ -1,16 +1,0 @@
-.home {
-  background-image: url("../img/showcase.jpg");
-  background-size: cover;
-  background-size: center;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  padding: 0 20px;
-}
-
-.title, .subtitle {
-  margin: 0
-}
