--- frontend/src/styles/readmore.css
@@ -1,14 +1,0 @@
-.read-btn {
-  font-size: 1.1em;
-  text-decoration: none;
-  color: black;
-  border: black 1px solid;
-  padding: 10px 20px;
-  border-radius: 10px;
-  margin-top: 10px;
-}
-
-.read-btn:hover {
-  background: black;
-  color: white;
-}
